Edit functionality created: Loads existing patient to form. Doesn't save yet
